Update Profile Img (pt. 2)
The ProfileImg model now stores image "names".
The user model will still have an ImageField where images will be uploaded to the 'images' directory in the project.
There is now a static method that will create new ProfileImg instances if there isn't one associated with a given user.
The route to update a user's profile image will pass an HTTP request and the file "name" for the new profile picture.