First pass of the mentor mfa system
This system is not connected to the login process yet.

To test replace str_recipient in mentor_mfa_request under under_development with a valid email address

Then start the server and navigate to localhost:8000/mentor/2fa

You will be redirected to localhost:8000/mentor/2fa/otp which will display a welcome message or an error message.